refactor(utils): move extractContentText to src/utils
